# pkg/api/docs/swagger.yaml
basePath: /
definitions:
...
  GitProvider:
    properties:
      baseApiUrl:
        type: string
      id:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
host: localhost:3986
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: 0.1.0
paths:
...
  /gitprovider:
    get:
      description: List Git providers
      operationId: ListGitProviders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/GitProvider'
            type: array
      summary: List Git providers
      tags:
      - gitProvider
    put:
      description: Set Git provider
      operationId: SetGitProvider
      parameters:
      - description: Git provider
        in: body
        name: gitProviderConfig
        required: true
        schema:
          $ref: '#/definitions/GitProvider'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Set Git provider
      tags:
      - gitProvider
  /gitprovider/{gitProviderId}:
    delete:
      description: Remove Git provider
      operationId: RemoveGitProvider
      parameters:
      - description: Git provider
        in: path
        name: gitProviderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Remove Git provider
      tags:
      - gitProvider
...
  /gitprovider/for-url/{url}:
    get:
      description: Get Git provider
      operationId: GetGitProviderForUrl
      parameters:
      - description: Url
        in: path
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GitProvider'
      summary: Get Git provider
      tags:
      - gitProvider
  /gitprovider/id-for-url/{url}:
    get:
      description: Get Git provider ID
      operationId: GetGitProviderIdForUrl
      parameters:
      - description: Url
        in: path
        name: url
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get Git provider ID
      tags:
      - gitProvider
...
schemes:
- http
security:
- Bearer: []
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"